+++ api/controllers/email.controller.js
@@ -1,3 +1,5 @@
+var nodemailer = require('nodemailer');
+
 /*
     Description
     Takes:
@@ -7,8 +9,29 @@
     }
     Calling route:
 */
-module.exports.sendEmail = function(req,res){
+module.exports.sendEmail = async function(email, subject, content){
+  try{
+    var transporter = nodemailer.createTransport({
+    service: 'gmail',
+    auth: {
+      user: 'pinmateapp@gmail.com',
+      pass: 'gradproject2018'
+    }
+    });
 
+    var mailOptions = {
+      from: 'pinmateapp@gmail.com',
+      to: email,
+      subject: subject,
+      text: content
+    };
+
+    var info = await transporter.sendMail(mailOptions);
+    return info;
+  }catch(e){
+    console.log(e);
+    throw e;
+  }
 }
 /*
     Description
