+++ package.json
@@ -18,6 +18,7 @@
     "mongoose-dummy": "^1.0.5",
     "morgan": "~1.9.0",
     "nodemailer": "^4.6.5",
+    "random-token": "0.0.8",
     "serve-favicon": "~2.4.5",
     "socket.io": "^2.1.1"
   },
